src/requirements.txt
aiojobs==1.0.0
alembic==1.8.0
amqp==5.1.1
anyio==3.6.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.2
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
cssutils==2.4.2
databases==0.6.0
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.14.1
email-validator==1.2.1
emails==0.6
environs==9.5.0
executing==0.8.3
fastapi==0.77.1
fastapi-contrib==0.2.11
fastapi-jsonrpc==2.2.2
fastapi-pagination==0.9.3
fastapi_auth_middleware==1.0.2
gunicorn==20.1.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
ipython==8.4.0
jedi==0.18.1
jsonrpcserver==5.0.7
jsonschema==4.6.0
kombu==5.2.4
lxml==4.9.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
matplotlib-inline==0.1.3
motor==3.0.0
OSlash==0.6.3
packaging==21.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
premailer==3.10.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
requests==2.28.0
rsa==4.8
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
SQLAlchemy-Utils==0.38.2
stack-data==0.3.0
starlette==0.19.1
tomli==2.0.1
traitlets==5.3.0
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
# https://pypi.org/project/us/